+++ packages/cardano-services/src/Program/defaults.ts
@@ -1,3 +1,9 @@
 export { OGMIOS_URL_DEFAULT, RABBITMQ_URL_DEFAULT } from '../ProgramsCommon';
 
+// http-server
 export const API_URL_DEFAULT = 'http://localhost:3000';
+
+// tx-worker
+export const PARALLEL_MODE_DEFAULT = false;
+export const PARALLEL_TXS_DEFAULT = 3;
+export const POLLING_CYCLE_DEFAULT = 500;
