+++ packages/cardano-services/src/Program/index.ts
@@ -1,6 +1,7 @@
 export * from './defaults';
 export * from './errors';
 export * from './loadHttpServer';
+export * from './loadTxWorker';
 export * from './ProgramOptionDescriptions';
 export * from './ServiceNames';
 export * from './utils';
