+++ packages/cardano-services/src/Program/loadTxWorker.ts
@@ -1,9 +1,8 @@
 import { CommonProgramOptions } from '../ProgramsCommon';
 import { Logger } from 'ts-log';
 import { TxSubmitWorkerConfig } from '../TxSubmit';
-import { createDnsResolver, getOgmiosTxSubmitProvider } from '../Program';
+import { createDnsResolver, getOgmiosTxSubmitProvider, getRunningTxSubmitWorker } from './';
 import { createLogger } from 'bunyan';
-import { getRunningTxSubmitWorker } from './utils';
 
 export type TxWorkerOptions = CommonProgramOptions &
   Pick<TxSubmitWorkerConfig, 'parallel' | 'parallelTxs' | 'pollingCycle'>;
