--- packages/cardano-services/src/TxWorker/TxWorkerOptionDescriptions.ts
@@ -1,5 +1,0 @@
-export enum TxWorkerOptionDescriptions {
-  Parallel = 'Parallel mode',
-  ParallelTxs = 'Parallel transactions',
-  PollingCycle = 'Polling cycle'
-}
