--- packages/cardano-services/src/TxWorker/defaults.ts
@@ -1,7 +1,0 @@
-export { OGMIOS_URL_DEFAULT, RABBITMQ_URL_DEFAULT } from '../ProgramsCommon';
-
-export const PARALLEL_MODE_DEFAULT = false;
-
-export const PARALLEL_TXS_DEFAULT = 3;
-
-export const POLLING_CYCLE_DEFAULT = 500;
