--- packages/cardano-services/src/TxWorker/index.ts
@@ -1,3 +1,0 @@
-export * from './defaults';
-export * from './loadTxWorker';
-export * from './TxWorkerOptionDescriptions';
