+++ packages/cardano-services/src/index.ts
@@ -4,7 +4,6 @@ export * from './Program';
 export * from './RunnableModule';
 export * from './StakePool';
 export * from './TxSubmit';
-export * from './TxWorker';
 export * from './Utxo';
 export * from './ChainHistory';
 export * as errors from './errors';
